Clean up changes
